Add base retrieval class
